Add simulated data from Gostic, 2020 for benchmarking (#37)
* Add simulated data from Gostic, 2020 for benchmarking

This commit re-uses the data processing and documentation from cdcgov/cfa-epinow2-pipeline#17. That repo is open source and public domain by nature of being USG property. I think it's convenient to re-use, but @athowes and @kgostic there's a little text in there that you both suggested, so I'd appreciate if you could give your permission for the re-use here! I've made sure to credit you both in the commit as it's partially your writing.

It's probably best practice to make the data prep and processing here fully reproducible in `data-raw/` but it's quite a pain to do, with a mixture of shells cripting and R scripting needed. I skipped it out of convenience, but if you have a really strong feeling that it's required @seabbs, let me know and we can revisit.

Closes #9

Co-authored-by: Adam Howes <adamthowes@gmail.com>
Co-authored-by: Katie Gostic <uep6@cdc.gov>

* Bump NEWS

* Coerce new `obs_incidence` col to integer from double

* Generate GI PMF in `data-raw/`

* Point to specific CFAEpiNow2Pipeline commit

* Remove copied & pasted unrelated text

* Re-render roxygen docs

* Add primarycensoreddist to remotes

* Typo

* Rename synthetic dataset and GI dist

* Pin primarycensorreddist to R-universe not github

---------

Co-authored-by: Adam Howes <adamthowes@gmail.com>
Co-authored-by: Katie Gostic <uep6@cdc.gov>