Set local merge strategy for NEWS (#58)
* Set local merge strategy for NEWS

The NEWS check is great because I always forget to update NEWS,
but the file also always causes merge conflicts because every PR
is changing the same line.

Here, I'm setting the merge strategy for NEWS.md only to be a
union of the changes in the two files. This gets rid of the merge
conflicts.

It's an imperfect solution because it only works locally. GitHub
can't use the merge strategy and will show the conflict. That's a
pain but a local solution is good enough for me for now.

* Bump NEWS