Auto-convert dates passed as strings (#67)
And print a helpful message when casting the dates. This is a small
change, but a nice quality-of-life improvement when specifying the
model.

This change breaks the pattern of checking & validating with empty
returns. I don't love breaking the pattern, but I thought the best
alternative would be to fully pull out the validators and have each
validator return their input. That might be worth it at some point,
but while we're only doing that for the dates I didn't think it was
worth it just yet.