Quick getting started vignette
We can add more here, but this walks through the basic functionality available now.
I tried to keep it pretty friendly and approachable with the idea that we can add more methods-focused content and guides for advanced users down the road.
I want to be able to hand the package and this vignette to a target user for user testing and see if they can handle it cold.

Closes #44