Add a day-of-week effect to simulated dataset
The current timeseries doesn't have a day of week effect, which makes it hard to test how the day of week effect implementation is performing.
This modifies the simulated timeseries to add a large-ish day of week reporting effect.

In implementing this change, I converted the timeseries over to using a `data-raw` script for reproducibility.
I moved the original Gostic "Practical considerations" dataset into an RDS object in `data-raw`.
This change makes the dataset tracked and available for re-generation of the package dataset, but doesn't expose it to or ship it with the package.
It's not 100% reproducible because I don't implement the git clone of the "Practical considerations" repo, but I think it gets us a fair but further than before.

I also clean up the package-provided `stochastic_sir_rt` datset.
I rename the `obs_incidence` col to `obs_cases`, add a `obs_cases_no_dow` for testing, and rename `date` to `reference_date` in anticipation of a future `report_date`.

Closes #90