Work around gratia bug in factor handling
Gratia development version works fine, but the current CRAN version has a bug in its factor handling.
When passing (known) factor levels, it errors out when constructing the dataset in `data_slic()`.

This error wasn't showing up in the tests because I had installed the development version from Github. But it was getting caught on CI and in the vignette because those were installing the latest CRAN version.

I had to implement an imperfect workaround here. It's messy because each individual factor level is "far" from the data in mgcv's definition -- at least far enough to throw a warning.

It would be pretty labor intensive to reconstruct the associated levels because we explicitly set them to zero with `exclude`. I don't want to do all the work of reconstructing the matching levels because to then immediately toss them. So instead, I've suppressed the warning -- an imperfect but "good enoough" solution...