Turn on dev site (#101)
* Turn on dev site

I _think_ from (these docs)[https://pkgdown.r-lib.org/articles/how-to-update-released-site.html] that this is enough?
I had some trouble following the bit at the end about cherry-picking stuff around, but it sounds like when `mode` is auto that it watches and updates appropriately based on version.

Let me know if this isn't right Sam and I'll give it a tweak.

Closes #99

* Bump NEWS