Allow users to add geoms to `plot()` return (#106)
* Allow users to add geoms to `plot()` return

In testing, I realized the current returned plot doesn't allow for
easy addition of additional geoms. The plot object as written takes
the data in the call to `ggplot()`. All subsequent geoms require the
data to contain the data in this default data object.

Because we're plotting draws, this is a pain. I've switched things
over to passing data explictly to each geom. This change removes the
default data specification, but doesn't change any of the plots
(see no change to the tests).

* Update Wordlist with new words

---------

Co-authored-by: zsusswein <zsusswein@users.noreply.github.com>