Add R setup to pre-commit workflow (#112)
* Add R setup to pre-commit workflow

In the last ~month, GitHub upgraded the default Ubuntu runner. In the
process they re-examined the default software on the runners, dropping
R.

This change re-adds R, fixing pre-commit runs.

* Bump NEWS

* Flyby fix: fix package name in rendered Rd

This Rd file is rendered as of #108 so is now throwing a warning
when I try to re-render the docs. Correct styling for packages is
explained here https://github.com/r-lib/roxygen2/issues/1578

* Bump pre-commit